feat: use unicode-range to avoid misusing font
This resolves a problem where the emoji font is used for *everything*,
which can't be fixed unless you specify a `font-family` property with an
specific font before it. This is a problem on Electron apps where you
want to continue using the default `sans-serif` meta-font, which never
takes precedence over the emoji font (even if it's specified earlier in
the font stack).

See also: https://github.com/fraction/emoji-unicode-range